pyproject.toml
[tool.poetry]
name = "sahmino"
version = "0.1.0"
description = "Backend for Sahmino (Django REST API)"
authors = ["Ali Arefi <arefi.ali90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.15"
drf-yasg = "^1.21"
drf-spectacular = "^0.28"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-django = "^4.5"
black = "^24.3"
isort = "^5.12"
mypy = "^1.7"
flake8 = "^7.0"
pre-commit = "^3.4"
flake8-pyproject = "^1.2.3"

[tool.flake8]
max-line-length = 91
extend-ignore = ["E203", "W503"]
exclude = [".venv", "migrations", "__pycache__"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".venv", "migrations", "__pycache__"]

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = ".venv|migrations|__pycache__"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = ".venv|migrations|__pycache__"

[tool.poetry.scripts]
# no CLI scripts for now

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
